Remove IDocumentProperties and its implementations.
Move FnDocumentProperties.getProperty and its helpers to
DocumentTraverser. The only changes are renaming the method to
getPropertyValues, and creating and returning the list instead of
taking it as a parameter.

Move the mock metadata from FileNetProxies.MockDocumentProperties
to ObjectMocks.mockDocument.
